move classes to shared
