+++ src/app/map/side-panel/side-panel.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, Output } from '@angular/core';
 import { AppStateService } from '../../services/app-state.service';
 import { attr } from '../../tree';
 import { trigger, state, style} from '@angular/animations';
-import { showSidePanel } from '../../classes/animations';
+import { showSidePanel } from '../../shared/animations';
 
 @Component({
   selector: 'side-panel',
