+++ src/app/shared/utils.ts
